requirements.txt
pillow
mopidy-asyncio-client
